requirements.txt
# ─── gRPC / Proto ─────────────────────────────────────────────────────────────
grpcio==1.65.1
grpcio-tools==1.65.1
protobuf==5.27.2

# ─── HTTP / Async ─────────────────────────────────────────────────────────────
aiohttp==3.9.5
asyncio==3.4.3

# ─── Data validation ──────────────────────────────────────────────────────────
pydantic==2.8.2

# ─── Gateway ──────────────────────────────────────────────────────────────────
fastapi==0.111.1
uvicorn[standard]==0.30.3
websockets==12.0

# ─── AI / RAG ─────────────────────────────────────────────────────────────────
google-generativeai==0.7.2
pinecone-client==4.1.0

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.3.1
pytest-asyncio==0.23.7
pytest-mock==3.14.0

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
